Implemented TCP networking clients and system-clock timers
The meter and utility network clients can now be implemented with
TcpNetworkClient and TcpUtilityClient, which use epoll() to multiplex
several incoming connections on a single socket without adding threads.
The TimerManager feature can now be imlemented with LinuxTimerManager,
which uses the system clock to set timeouts, which is necessary in order
to use the TCP networking clients (since they need to use real time to
measure timeouts, not event simulator clock ticks).

Also, I changed the build system to use an explicit Makefile instead of
relying on Eclipse's CDT builder to figure it out, because now there are
multiple possible "main" files depending on whether we wish to compile
in simulated mode or in "real networking" mode.