Fixed lots of bugs in networking
There were several small, nasty bugs in my implementations of
serialization functions and the network protocol I wrote in
TcpNetworkClient. I think they have all been fixed now, and the
end-to-end test EmulatedTestMain.cpp works on 11 nodes.

LinuxTimerManager was also buggy, because it had a race condition
between timers firing and being cancelled. The workaround for now is to
not delete cancelled timers.