Rewrote networking to use long-lived sockets
Instead of opening a new socket for every message, and using "connection
closed" to indicate the end of a message, clients now keep sockets open
indefinitely and prefix every message with its size in bytes. Also
updated mutils and updated Makefile to require C++17 because the new
version of mutils-serialization no longer works with C++14.