Replaced LinuxTimerManager with non-buggy version
The design of LinuxTimerManager was broken, because it turns out you
can't safely do nontrivial things (like std::map::erase) inside a signal
handler. Rewriting it to pass fired-timer IDs to a dedicated handler
thread required bringing in another new library, which contains an
implementation of a lock-free queue (locks are also not allowed in a
signal handler).